requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# YOLOv8 and tracking
ultralytics==8.0.200
lap==0.4.0  # For ByteTrack
scipy==1.11.3
scikit-learn==1.3.1

# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Database
sqlalchemy==2.0.21
alembic==1.12.0

# Notifications
requests==2.31.0
python-dotenv==1.0.0

# Video processing
imageio==2.31.5
imageio-ffmpeg==0.4.9

# Utilities
pyyaml==6.0.1
click==8.1.7
schedule==1.2.0
python-dateutil==2.8.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1